% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_raw_data.R
\name{cleanStormData}
\alias{cleanStormData}
\title{Reformat storm data}
\usage{
cleanStormData(data)
}
\arguments{
\item{data}{input data from fwf imported using \code{parseStormData}}
}
\value{
data frame with tidy data
}
\description{
Changes format for `storm_id`, transforms longitude for values in Western Hemisphere,
combines columns describing data and time to create a single variable for each
observation, and converts data into "long" format with separate rows for each of the
three wind speeds for wind radii (34 kts, 50 kts, and 64 kts).
}
