% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurricane.R
\docType{data}
\name{GeomHurricane}
\alias{GeomHurricane}
\title{Function to construct a new class for the geom_hurricane, this function will create a wind radii for a
given storm obseravtion. It is to be used with maps construct with get_map}
\format{An object of class \code{GeomHurricane} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomHurricane
}
\arguments{
\item{required_aes}{required aesthetic arguments for the geom_hurricane supplied in character vector}

\item{default_aes}{default values for aesthetic arguments}

\item{draw_key}{the function to draw the legend with the associated geom}

\item{draw_group}{where the bulk of this geom is constructed}
}
\description{
Function to construct a new class for the geom_hurricane, this function will create a wind radii for a
given storm obseravtion. It is to be used with maps construct with get_map
}
\examples{
\dontrun{
  geom_hurricane(data = storm_observation, aes(x = longitude, y = latitude, r_ne = ne, r_se = se, r_nw =
  nw, r_sw = sw, fill = wind_speed, color = wind_speed)
}
}
\keyword{datasets}
