% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurricane.R
\name{mapObservation}
\alias{mapObservation}
\title{Map an observation of a storm}
\usage{
mapObservation(obs)
}
\arguments{
\item{obs}{an observation retrieved from \code{stormObservation}}
}
\description{
Map an observation of a storm
}
