% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurricane.R
\docType{data}
\name{StatHurricane}
\alias{StatHurricane}
\title{Constructs a hurricane wind radii stat}
\format{An object of class \code{StatHurricane} (inherits from \code{Stat}, \code{ggproto}) of length 3.}
\usage{
StatHurricane
}
\arguments{
\item{required_aes}{aesthetics required for computing polygon coordinates}

\item{compute_group}{function to calculate polygon for each observation group}
}
\value{
data frame with x,y coordinates for plotting wind radii
}
\description{
Performs distance calculations, transforming data into appropriate format for plotting
as wind radii polygons
}
\keyword{datasets}
